Flesh out what i currently think is the complete list of v1 tutorials.
Add crappy breadcrumb navigation to the sidebar.  Needs some css love, but it works.

Move the playground widget out to its own directory so it doesn't show
up in the tutorial navigation.  Need to figure out how to include one
or more of these widgets in a tutorial.

Still need numbers in the file names to get the right ordering via
haiku.  Added sort fields pending a server fix, then we can drop the
numbers from the file names to make the URL's less ugly.