Use the "roaming" instead of "generic" runtime factory.
See vanadium/issues#1289

This required an unfortunate hack to the example fortune server in the
tests here so that the test continues to pass when run on Google Compute
Engine virtual machines.

A more appropriate fix: Setting a tight ListenSpec when running code on
the playground is left as an excercise for the future.

Change-Id: Ie7787d7d0264706d20b7289d06c6523abb6a695c